requirements.txt
numpy 1.24.2
matplotlib 3.63.3
scikit-image 0.19.3
scikit-learn 1.2.2
reedsolo 1.7.0
pyzbar 0.1.9
pandas 2.0.1
pillow 9.4.0